Some trouble-shooting; fiddled with almost every class

+++ src/org/academiadecodigo/haltistas/besquare/client/GameField.java
@@ -2,6 +2,7 @@ package org.academiadecodigo.haltistas.besquare.client;
 
 import org.academiadecodigo.haltistas.besquare.FilePath;
 import org.academiadecodigo.simplegraphics.pictures.Picture;
+
 import java.util.Map;
 
 public class GameField {
@@ -14,13 +15,13 @@ public class GameField {
     private CharacterSprite p2Sprite;
 
 
-    GameField(){
+    GameField() {
         p1Sprite = new CharacterSprite(FilePath.PLAYER1);
         p2Sprite = new CharacterSprite(FilePath.PLAYER2);
     }
 
 
-    protected void moveCharacters(int player1X, int player1Y, int player2X, int player2Y){ // TODO: resolv temporary solution
+    protected void moveCharacters(int player1X, int player1Y, int player2X, int player2Y) { // TODO: resolv temporary solution
 
         int xTranslation = ((p1Sprite.getSprite().getX() - player1X) + GameField.PADDING);
         int yTranslation = ((p1Sprite.getSprite().getY() - player1Y) + GameField.PADDING);
@@ -33,15 +34,18 @@ public class GameField {
         p2Sprite.getSprite().translate(xTranslation, yTranslation);
     }
 
-    protected void loadBackground(String path){
+    protected void loadBackground(String path) {
 
-        background.delete();
+        if (background != null) {
+            background.delete();
+        }
+
         background = new Picture(PADDING, PADDING, path);
         background.draw();
     }
 
 
-    protected void loadCharacters(int player1X, int player1Y, int player2X, int player2Y){
+    protected void loadCharacters(int player1X, int player1Y, int player2X, int player2Y) {
 
         p1Sprite.getSprite().delete();
         p2Sprite.getSprite().delete();
@@ -52,6 +56,7 @@ public class GameField {
         p1Sprite.getSprite().draw();
         p2Sprite.getSprite().draw();
 
+        System.out.println("loaded characters");
     }
 }
